requirements.txt
pypdf==3.17.4
Pillow==10.1.0
pdf2image==1.17.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==24.1.0
requests